pyproject.toml
[project]
name = "skill-security-scanner"
version = "0.1.0"
description = "一个用于检测 Claude Code Skills 安全性的工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "colorama>=0.4.6",
    "tree-sitter>=0.20.0",
    "tree-sitter-bash>=0.20.0",
    "tree-sitter-python>=0.20.0",
    "tree-sitter-javascript>=0.20.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
]
